% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchor_pair_enrich.R
\name{anchor_pair_enrich}
\alias{anchor_pair_enrich}
\title{Determine enriched motifs in anchors}
\usage{
anchor_pair_enrich(
  interaction_data,
  method = c("countCorrelation", "scoreCorrelation", "countHypergeom", "countFisher")
)
}
\arguments{
\item{interaction_data}{an interactionData object of paired genomic regions}

\item{method}{choice of method for co-occurrence include
\code{countCorrelation}, \code{scoreCorrelation}, \code{countHypergeom}, or
\code{countFisher}}
}
\value{
an interactionData object where \code{obj$pair_motif_enrich} contains
the p-values for significance of seeing a higher co-occurrence than
what we get by chance.
}
\description{
Determine whether motifs between paired bed regions have a statistically
significant relationship. Options for significance are motif score
correlation, motif count correlation, or hypergeometric motif co-occurrence.
}
\examples{
genome <- BSgenome.Mmusculus.UCSC.mm9::BSgenome.Mmusculus.UCSC.mm9
motif_file <- system.file(
  "extdata/consensus_HOCOMOCOv11_core_MOUSE-plus_YY1.piq",
  package = "spatzie")
motifs <- TFBSTools::readJASPARMatrix(motif_file, matrixClass = "PFM")

yy1_pd_interaction <- scan_motifs(spatzie:::yy1_interactions, motifs, genome)
yy1_pd_interaction <- filter_motifs(yy1_pd_interaction, 0.4)
yy1_pd_count_corr <- anchor_pair_enrich(yy1_pd_interaction,
                                        method = "countCorrelation")
}
\author{
Jennifer Hammelman
}
