% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anchor_pair_example_count}
\alias{anchor_pair_example_count}
\title{spatzie count correlation data set}
\format{
An interactionData object
}
\usage{
data(anchor_pair_example_count)
}
\description{
This object contains genomic interactions obtained by mouse YY1 ChIA-PET
scanned for mouse transcription factor motifs, filtered for motifs present
in at least 10% of interactions, and processed for significant motif:motif
interactions with count correlation. It serves as unit test data.
}
\keyword{datasets}
