% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{compare_pairs_example}
\alias{compare_pairs_example}
\title{compare_motif_pairs example}
\format{
A matrix
}
\usage{
data(compare_pairs_example)
}
\description{
This is a matrix containing example result from compare_motif_pairs. It
serves as unit test data.
}
\keyword{datasets}
