% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_motifs.R
\name{filter_motifs}
\alias{filter_motifs}
\title{Filter motifs based on occurrence within interaction data}
\usage{
filter_motifs(interaction_data, threshold)
}
\arguments{
\item{interaction_data}{an interactionData object of paired genomic regions}

\item{threshold}{fraction of interactions that should contain a motif for a
motif to be considered}
}
\value{
an interactionData object where \code{obj$anchor1_motif_indices}
and \code{obj$anchor2_motif_indices} have been filtered to motifs that are
present in a threshold fraction of interactions
}
\description{
Select a subset of motifs that are in at least a threshold fraction of
regions. Motif subsets are selected separately for anchor one and anchor
two regions.
}
\examples{
# TODO

}
\author{
Jennifer Hammelman
}
