% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{filter_pairs_example}
\alias{filter_pairs_example}
\title{spatzie score correlation filtered data set}
\format{
A interactionData object
}
\usage{
data(filter_pairs_example)
}
\description{
This object contains genomic interactions obtained by mouse YY1 ChIA-PET
scanned for mouse transcription factor motifs, filtered for motifs present
in at least 10% of interactions, processed for significant motif:motif
interactions with score correlation, and filtered for pairs with p < 0.5.
It serves as unit test data.
}
\keyword{datasets}
