% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ep_coenrichment.R
\name{find_ep_coenrichment}
\alias{find_ep_coenrichment}
\title{Find co-enriched motif pairs in enhancer-promoter interactions}
\usage{
find_ep_coenrichment(
  int_raw_data,
  motifs_file,
  motifs_file_matrix_format = c("pfm", "ppm", "pwm"),
  genome_id = c("hg38", "hg19", "mm9", "mm10"),
  cooccurrence_method = c("countCorrelation", "scoreCorrelation", "countHypergeom",
    "countFisher"),
  filter_threshold = 0.4
)
}
\arguments{
\item{int_raw_data}{a data frame with at least six columns:
\tabular{rl}{
  column 1: \tab character; genomic location of interaction anchor 1 -
  chromosome (e.g., \code{"chr3"})\cr
  column 2: \tab integer; genomic location of interaction anchor 1 -
  start coordinate\cr
  column 3: \tab integer; genomic location of interaction anchor 1 -
  end coordinate\cr
  column 4: \tab character; genomic location of interaction anchor 2 -
  chromosome (e.g., \code{"chr3"})\cr
  column 5: \tab integer; genomic location of interaction anchor 2 -
  start coordinate\cr
  column 6: \tab integer; genomic location of interaction anchor 2 -
  end coordinate
}}

\item{motifs_file}{\href{http://jaspar.genereg.net/faq/}{JASPAR format}
matrix file containing multiple motifs to scan for}

\item{motifs_file_matrix_format}{type of position-specific scoring matrices
in \code{motifs_file}, valid options include:
\tabular{rl}{
  \code{pfm}: \tab position frequency matrix, elements are absolute
  frequencies, i.e., counts (default)\cr
  \code{ppm}: \tab position probability matrix, elements are probabilities,
  i.e., Laplace smoothing corrected relative frequencies\cr
  \code{pwm}: \tab position weight matrix, elements are log likelihoods
}}

\item{genome_id}{ID of genome assembly interactions in \code{int_raw_data}
were aligned to, valid options include \code{hg19}, \code{hg38}, \code{mm9},
and \code{mm10}, defaults to \code{hg38}}

\item{cooccurrence_method}{choice of method for co-occurrence include
\code{countCorrelation}, \code{scoreCorrelation}, \code{countHypergeom}, or
\code{countFisher}, see \code{\link{anchor_pair_enrich}}, defaults to
\code{countCorrelation}}

\item{filter_threshold}{fraction of interactions that should contain a
motif for a motif to be considered, see \code{\link{filter_motifs}},
defaults to \code{0.4}}
}
\value{
a list with the following items:
\tabular{rl}{
  \code{int_data} \tab
  \code{\link[GenomicInteractions]{GenomicInteractions}} object;
  promoter-enhancer interactions\cr
 \code{int_data_motifs}: \tab \code{interactionData} object; return value of
 \code{\link{scan_motifs}}\cr
  \code{filtered_int_data_motifs}: \tab \code{interactionData} object;
  return value of \code{\link{filter_motifs}}\cr
  \code{annotation_pie_chart}: \tab ggplot2 plot; return value of
  \code{\link[GenomicInteractions]{plotInteractionAnnotations}}\cr
  \code{motif_cooccurrence}: \tab \code{interactionData} object; return
  value of \code{\link{anchor_pair_enrich}}
}
}
\description{
Identifies co-enriched pairs of motifs in enhancer-promoter interactions
selected from a data frame of general genomic interactions.

Calls functions \code{\link{scan_motifs}}, \code{\link{filter_motifs}},
and \code{\link{anchor_pair_enrich}} internally.

Uses \code{biomaRt} to retrieve annotations.
}
\examples{
\dontrun{
interactions_file <- system.file("extdata/yy1_interactions.bedpe",
                                 package = "spatzie")
motifs_file <- system.file(
  "extdata/consensus_HOCOMOCOv11_core_MOUSE-plus_YY1.piq",
  package = "spatzie")

df <- read.table(interactions_file, header = FALSE, sep = "\t")
res <- find_ep_coenrichment(df, motifs_file,
                            motifs_file_matrix_format = "pfm",
                            genome_id = "mm10")
}

}
\author{
Jennifer Hammelman

Konstantin Krismer
}
