% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpecificInteractions.R
\name{get_specific_interactions}
\alias{get_specific_interactions}
\title{Get interactions that contain a specific motif pair}
\usage{
get_specific_interactions(
  interactionData,
  anchorOneMotif = "",
  anchorTwoMotif = ""
)
}
\arguments{
\item{interactionData}{TODO}

\item{anchorOneMotif}{TODO}

\item{anchorTwoMotif}{TODO}
}
\value{
TODO
}
\description{
Select interactions that contain anchorOneMotif within anchorOne and
anchorTwoMotif within anchorTwo.
}
\examples{
# TODO

}
\author{
Jennifer Hammelman
}
