% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{int_data_mslcl}
\alias{int_data_mslcl}
\title{MSLCL Enhancer - Promoter Interactions Data Set}
\format{
An \code{interactionData} object
}
\usage{
data(int_data_mslcl)
}
\description{
This object contains genomic interactions obtained by human RAD21 ChIA-PET
from MSLCL cells and serves as unit test data.
}
\keyword{datasets}
