% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{int_data_yy1}
\alias{int_data_yy1}
\title{Mouse YY1 Enhancer - Promoter Interactions Data Set}
\format{
An \code{interactionData} object
}
\usage{
data(int_data_yy1)
}
\description{
This object contains genomic interactions obtained by mouse YY1 ChIA-PET
and serves as example and unit test data.
}
\keyword{datasets}
