% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{interactionDataK562}
\alias{interactionDataK562}
\title{K562 Enhancer - Promoter Interactions Data Set}
\format{
A \code{\link[GenomicInteractions]{GenomicInteractions}} object
}
\usage{
data(interactionDataK562)
}
\description{
This object contains genomic interactions obtained by human RAD21 ChIA-PET
from K562 cells and serves as unit test data.
}
\keyword{datasets}
