% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{interactionDataMSLCL}
\alias{interactionDataMSLCL}
\title{MSLCL Enhancer - Promoter Interactions Data Set}
\format{
A \code{\link[GenomicInteractions]{GenomicInteractions}} object
}
\usage{
data(interactionDataMSLCL)
}
\description{
This object contains genomic interactions obtained by human RAD21 ChIA-PET
from MSLCL cells and serves as unit test data.
}
\keyword{datasets}
