% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{interactions_yy1}
\alias{interactions_yy1}
\title{Mouse YY1 Enhancer - Promoter Interactions Data Set}
\format{
A \code{\link[GenomicInteractions]{GenomicInteractions}} object
}
\usage{
data(interactions_yy1)
}
\description{
This object contains genomic interactions obtained by mouse YY1 ChIA-PET
and serves as example and unit test data. The same data set is used in
the vignette.
}
\keyword{datasets}
