% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{interactions_yy1_enhancer}
\alias{interactions_yy1_enhancer}
\title{Mouse YY1 Enhancer - Promoter Interactions Data Set - YY1 enhancers}
\format{
A \code{\link[GenomicInteractions]{GenomicInteractions}} object
}
\usage{
data(interactions_yy1_enhancer)
}
\description{
This is a \code{\link[GenomicInteractions]{GenomicInteractions}} object
containing proccessed results from YY1
ChIA-PET of interactions that contain a YY1 motif in the enhancer
(anchor 2) region. It serves as unit test data.
}
\keyword{datasets}
