% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_enrich.R
\name{overlap_enrich}
\alias{overlap_enrich}
\title{TODO}
\usage{
overlap_enrich(
  interactions,
  all_interactions,
  bedranges,
  anchor = c("anchorOne", "anchorTwo")
)
}
\arguments{
\item{interactions}{TODO}

\item{all_interactions}{TODO}

\item{bedranges}{TODO}

\item{anchor}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO

}
\author{
Jennifer Hammelman
}
