% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapEnrich.R
\name{overlapEnrich}
\alias{overlapEnrich}
\title{TODO}
\usage{
overlapEnrich(
  interactions,
  allInteractions,
  bedranges,
  anchor = c("anchorOne", "anchorTwo")
)
}
\arguments{
\item{interactions}{TODO}

\item{allInteractions}{TODO}

\item{bedranges}{TODO}

\item{anchor}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\author{
Jennifer Hammelman
}
