% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_differential_interactions.R
\name{plot_differential_interactions}
\alias{plot_differential_interactions}
\title{Plot differential interactions}
\usage{
plot_differential_interactions(compared_motif_pairs)
}
\arguments{
\item{compared_motif_pairs}{TODO}
}
\value{
TODO
}
\description{
Plots clustered heatmap of log likelihood ration for motif interactions
between two datasets
}
\examples{
# TODO

}
\author{
Jennifer Hammelman
}
