% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEmbeddedInteractions.R
\name{plot_embedded_interactions}
\alias{plot_embedded_interactions}
\title{Plot embedded interactions}
\usage{
plot_embedded_interactions(interactions, color_by_motif, seed = 9302020)
}
\arguments{
\item{interactions}{TODO}

\item{color_by_motif}{TODO}

\item{seed}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO

}
\author{
Jennifer Hammelman
}
