% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_motif_occurrence.R
\name{plot_motif_occurrence}
\alias{plot_motif_occurrence}
\title{Plot motif occurrence}
\usage{
plot_motif_occurrence(
  interaction_data,
  method = c("counts", "matches", "scores")
)
}
\arguments{
\item{interaction_data}{TODO}

\item{method}{TODO}
}
\value{
TODO
}
\description{
Plots a histogram of motif values (either counts, instances, or scores)
for anchor 1 and anchor 2 regions.
}
\examples{
# TODO

}
\author{
Jennifer Hammelman
}
