% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_motif_pairs_heatmap.R
\name{plot_motif_pairs_heatmap}
\alias{plot_motif_pairs_heatmap}
\title{Plot motif pair occurrence}
\usage{
plot_motif_pairs_heatmap(pairEnrichmentMatrix)
}
\arguments{
\item{pairEnrichmentMatrix}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO

}
\author{
Jennifer Hammelman
}
