% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDifferentialInteractions.R
\name{plotDifferentialInteractions}
\alias{plotDifferentialInteractions}
\title{Plot differential interactions}
\usage{
plotDifferentialInteractions(comparedMotifPairs)
}
\arguments{
\item{comparedMotifPairs}{TODO}
}
\value{
TODO
}
\description{
Plots clustered heatmap of log likelihood ration for motif interactions
between two datasets
}
\author{
Jennifer Hammelman
}
