% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEmbeddedInteractions.R
\name{plotEmbeddedInteractions}
\alias{plotEmbeddedInteractions}
\title{Plot embedded interactions}
\usage{
plotEmbeddedInteractions(interactions, color_by_motif, seed = 9302020)
}
\arguments{
\item{interactions}{TODO}

\item{color_by_motif}{TODO}

\item{seed}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\author{
Jennifer Hammelman
}
