% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMotifOccurrence.R
\name{plotMotifOccurrence}
\alias{plotMotifOccurrence}
\title{Plot motif occurrence}
\usage{
plotMotifOccurrence(interactionData, method = c("counts", "matches", "scores"))
}
\arguments{
\item{interactionData}{TODO}

\item{method}{TODO}
}
\value{
TODO
}
\description{
Plots a histogram of motif values (either counts, instances, or scores)
for anchorOne and anchorTwo regions.
}
\author{
Jennifer Hammelman
}
