% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMotifPairEnrichment.R
\name{plotMotifPairsHeatmap}
\alias{plotMotifPairsHeatmap}
\title{Plot motif pair occurrence}
\usage{
plotMotifPairsHeatmap(pairEnrichmentMatrix)
}
\arguments{
\item{pairEnrichmentMatrix}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\author{
Jennifer Hammelman
}
