% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scan_interactions_example}
\alias{scan_interactions_example}
\title{Interactions scanned for motifs - interactionData object}
\format{
An interactionData object
}
\usage{
data(scan_interactions_example)
}
\description{
This object contains genomic interactions obtained by mouse YY1 ChIA-PET
scanned for mouse transcription factor motifs and serves as unit test data.
}
\keyword{datasets}
