% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scan_interactions_example_filtered}
\alias{scan_interactions_example_filtered}
\title{Interactions with motifs filtered for significance - interactionData object}
\format{
An interactionData object
}
\usage{
data(scan_interactions_example_filtered)
}
\description{
This object contains genomic interactions obtained by mouse YY1 ChIA-PET
scanned for mouse transcription factor motifs and filtered for motifs present
in at least 10% of interactions. It serves as unit test data.
}
\keyword{datasets}
