% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_motifs.R
\name{scan_motifs}
\alias{scan_motifs}
\title{Scans interaction file for motif instances}
\usage{
scan_motifs(interactions, motifs, genome)
}
\arguments{
\item{interactions}{an interactionData object of paired genomic regions}

\item{motifs}{a TFBS tools matrix of DNA binding motifs}

\item{genome}{a Biostrings genome must match chromosomes from interaction
data file}
}
\value{
an interaction data object where \code{obj$anchor1_motifs} and
\code{obj$anchor2_motifs} contain information about the scores and matches
to motifs from anchor one and anchor two of interaction data genomic regions
}
\description{
Uses motifmatchR to scan interaction regions for given motifs.
}
\examples{
genome <- BSgenome::getBSgenome("BSgenome.Mmusculus.UCSC.mm9")
motifs_file <- system.file("extdata/motifs_subset.txt.gz",
                           package = "spatzie")
motifs <- TFBSTools::readJASPARMatrix(motifs_file, matrixClass = "PFM")

yy1_pd_interaction <- scan_motifs(spatzie::interactions_yy1, motifs, genome)

}
\author{
Jennifer Hammelman
}
