% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanMotifs.R
\name{scanMotifs}
\alias{scanMotifs}
\title{Scans interaction file for motif instances}
\usage{
scanMotifs(interactions, motifs, genome)
}
\arguments{
\item{interactions}{an interactionData object of paired genomic regions}

\item{motifs}{a TFBS tools matrix of DNA binding motifs}

\item{genome}{a Biostrings genome must match chromosomes from interaction
data file}
}
\value{
an interaction data object where \code{obj$anchorOneMotifs} and
\code{obj$anchorTwoMotifs} contain information about the scores and matches
to motifs from anchor one and anchor two of interaction data genomic regions
}
\description{
Uses motifmatchR to scan interaction regions for given motifs.
}
\author{
Jennifer Hammelman
}
