% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{spatzie}
\alias{spatzie}
\title{spatzie}
\description{
Looks for motifs which are significantly co-enriched from enhancer-promoter
interaction data, derived from assays such as as HiC, ChIA-PET, etc. It can
also look for differentially enriched motif pairs between to interaction
experiments.
}
\author{
Jennifer Hammelman

Konstantin Krismer
}
