% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yy1_pd_interaction}
\alias{yy1_pd_interaction}
\title{yy1_pd_interaction}
\format{
An interactionData object
}
\usage{
data(yy1_pd_interaction)
}
\description{
This is a interactionData object containing processed results from YY1 ChIA-PET
interaction data. It serves as unit test data.
}
\keyword{datasets}
