% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yy1P_interactions}
\alias{yy1P_interactions}
\title{yy1P_interactions}
\format{
A GenomicInteractions object
}
\usage{
data(yy1P_interactions)
}
\description{
This is a GenomicInteractions object containing proccessed results from YY1 ChIA-PET
of interactions that contain a YY1 motif in the promoter (anchor 1) region. It serves
as unit test data.
}
\keyword{datasets}
