% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yy1p_yy1e_interactions}
\alias{yy1p_yy1e_interactions}
\title{yy1p_yy1e_interactions}
\format{
A \code{\link[GenomicInteractions]{GenomicInteractions}} object
}
\usage{
data(yy1p_yy1e_interactions)
}
\description{
This is a \code{\link[GenomicInteractions]{GenomicInteractions}} object
containing proccessed results from YY1
ChIA-PET of interactions that contain a YY1 motif in the promoter
(anchor 1) region and a YY1 motif in the enhancer (anchor 2) region.
It serves as unit test data.
}
\keyword{datasets}
