% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchorPairEnrich.R
\name{anchor_pair_enrich}
\alias{anchor_pair_enrich}
\title{Determine enriched motifs in anchors}
\usage{
anchor_pair_enrich(
  interactionData,
  method = c("countCorrelation", "scoreCorrelation", "countHypergeom", "countFisher")
)
}
\arguments{
\item{interactionData}{an interactionData object of paired genomic regions}

\item{method}{choice of method for co-occurrence include
\code{countCorrelation}, \code{scoreCorrelation}, \code{countHypergeom}, or
\code{countFisher}}
}
\value{
an interactionData object where \code{obj$pairMotifEnrich} contains
the p-values for significance of seeing a higher co-occurrence than
what we get by chance.
}
\description{
Determine whether motifs between paired bed regions have a statistically
significant relationship. Options for significance are motif score
correlation, motif count correlation, or hypergeometric motif co-occurrence.
}
\examples{
# TODO

}
\author{
Jennifer Hammelman
}
