% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareMotifPairs.R
\name{compare_motif_pairs}
\alias{compare_motif_pairs}
\title{Compare pairs of motifs between two interaction datasets}
\usage{
compare_motif_pairs(interactionData1, interactionData2, differential_p = 0.05)
}
\arguments{
\item{interactionData1}{TODO}

\item{interactionData2}{TODO}

\item{differential_p}{TODO}
}
\value{
TODO
}
\description{
Compute the log-likelihood ratio that a motif pair is differential between
two interaction datasets. Note that motif pair significance should have
been computed using the same method for both datasets.
}
\examples{
# TODO

}
\author{
Jennifer Hammelman
}
