% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterPairMotifs.R
\name{filter_pair_motifs}
\alias{filter_pair_motifs}
\title{Filter significant motif interactions}
\usage{
filter_pair_motifs(
  interactionData,
  method = p.adjust.methods,
  threshold = 0.05
)
}
\arguments{
\item{interactionData}{TODO}

\item{method}{TODO}

\item{threshold}{TODO}
}
\value{
TODO
}
\description{
Multiple hypothesis correction applied to filter for significant motif
interactions.
}
\examples{
# TODO

}
\author{
Jennifer Hammelman
}
