% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_specific_interactions.R
\name{get_specific_interactions}
\alias{get_specific_interactions}
\title{Get interactions that contain a specific motif pair}
\usage{
get_specific_interactions(
  interaction_data,
  anchor1_motif = NULL,
  anchor2_motif = NULL
)
}
\arguments{
\item{interaction_data}{TODO}

\item{anchor1_motif}{TODO}

\item{anchor2_motif}{TODO}
}
\value{
a subset of interactions that are only contain \code{anchor1_motif}
in anchor 1 and \code{anchor2_motif} in anchor 2
}
\description{
Select interactions that contain anchor1_motif within anchor 1 and
anchor2_motif within anchor 2.
}
\examples{
# TODO

}
\author{
Jennifer Hammelman
}
