% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_specific_interactions.R
\name{get_specific_interactions}
\alias{get_specific_interactions}
\title{Get interactions that contain a specific motif pair}
\usage{
get_specific_interactions(
  interaction_data,
  anchor1_motif = NULL,
  anchor2_motif = NULL
)
}
\arguments{
\item{interaction_data}{an interactionData object of paired genomic regions}

\item{anchor1_motif}{Motif name from \code{interactionData$anchor1_motifs}}

\item{anchor2_motif}{Motif name from \code{interactionData$anchor2_motifs}}
}
\value{
a \code{\link[GenomicInteractions]{GenomicInteractions}} object
containing a subset subset of interactions that contain an instance of
\code{anchor1_motif} in anchor 1 and \code{anchor2_motif} in anchor 2
}
\description{
Select interactions that contain anchor1_motif within anchor 1 and
anchor2_motif within anchor 2.
}
\examples{
genome <- BSgenome.Mmusculus.UCSC.mm9::BSgenome.Mmusculus.UCSC.mm9
motif_file <- system.file(
  "extdata/consensus_HOCOMOCOv11_core_MOUSE-plus_YY1.piq",
  package = "spatzie")
motifs <- TFBSTools::readJASPARMatrix(motif_file, matrixClass = "PFM")

yy1_pd_interaction <- scan_motifs(spatzie:::yy1_interactions, motifs, genome)
yy1_pd_interaction <- filter_motifs(yy1_pd_interaction, 0.4)
yy1_pd_count_corr <- anchor_pair_enrich(yy1_pd_interaction,
                                        method = "scoreCorrelation")
yy1_yy1_interactions <- get_specific_interactions(
  yy1_pd_interaction,
  anchor1_motif = "TYY1_MOUSE.H11MO.0.A",
  anchor2_motif = "TYY1_MOUSE.H11MO.0.A")

}
\author{
Jennifer Hammelman
}
