% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMotifOccurrence.R
\name{plot_motif_occurrence}
\alias{plot_motif_occurrence}
\title{Plot motif occurrence}
\usage{
plot_motif_occurrence(
  interactionData,
  method = c("counts", "matches", "scores")
)
}
\arguments{
\item{interactionData}{TODO}

\item{method}{TODO}
}
\value{
TODO
}
\description{
Plots a histogram of motif values (either counts, instances, or scores)
for anchorOne and anchorTwo regions.
}
\examples{
# TODO

}
\author{
Jennifer Hammelman
}
