% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_motif_pairs_heatmap.R
\name{plot_motif_pairs_heatmap}
\alias{plot_motif_pairs_heatmap}
\title{Plot motif pair occurrence}
\usage{
plot_motif_pairs_heatmap(pair_enrichment_matrix)
}
\arguments{
\item{pair_enrichment_matrix}{matrix of format \code{obj$pair_motif_enrich}
or \code{obj$pair_motif_enrich_sig}}
}
\value{
heatmap of input data
}
\description{
Creates heatmap to plot values associated with interaction pairs
}
\examples{
genome <- BSgenome.Mmusculus.UCSC.mm9::BSgenome.Mmusculus.UCSC.mm9
motif_file <- system.file(
  "extdata/consensus_HOCOMOCOv11_core_MOUSE-plus_YY1.piq",
  package = "spatzie")
motifs <- TFBSTools::readJASPARMatrix(motif_file, matrixClass = "PFM")

yy1_pd_interaction <- scan_motifs(spatzie:::yy1_interactions, motifs, genome)
yy1_pd_interaction <- filter_motifs(yy1_pd_interaction, 0.4)
yy1_pd_score_corr <- anchor_pair_enrich(yy1_pd_interaction,
                                        method = "scoreCorrelation")
plot_motif_pairs_heatmap(-log2(yy1_pd_score_corr$pair_motif_enrich))

}
\author{
Jennifer Hammelman
}
